Fixes  Issue 431 , added predefined classical equation (SIR)
git-svn-id: https://gama-platform.googlecode.com/svn/branches/GAMA_CURRENT@7164 b91ef52e-2b48-11df-9ab6-b1cdc0d4a89a

+++ ummisco.gaml.extensions.maths/src/ummisco/gaml/extensions/maths/utils/Solver.java
@@ -13,6 +13,6 @@ public abstract class Solver {
 	// Call the integrator, which should call computeDerivatives on the system
 	// of equations;
 	public abstract void solve(IScope scope, SystemOfEquationsStatement eq,
-			double time_initial, double time_final, int cycle_length);
+			double time_initial, double time_final, double cycle_length);
 
 }
